feat(tooltip): add tooltip input for showing plain string #10100
