+++ projects/igniteui-angular/src/lib/directives/tooltip/index.ts
@@ -1,0 +1,3 @@
+export * from './tooltip.directive';
+export * from './tooltip-target.directive';
+export * from './tooltip.module';
