+++ projects/igniteui-angular/src/lib/directives/tooltip/tooltip.component.html
@@ -1,0 +1,1 @@
+<span igxTooltip>{{content}}</span>
