+++ projects/igniteui-angular/src/lib/test-utils/tooltip-components.spec.ts
@@ -2,7 +2,7 @@ import { Component, ViewChild } from '@angular/core';
 import {
     IgxTooltipTargetDirective, IgxTooltipDirective,
     ITooltipShowEventArgs, ITooltipHideEventArgs
-} from '../directives/tooltip/tooltip.directive';
+} from '../directives/tooltip';
 
 @Component({
     template: `
