+++ projects/igniteui-angular/src/public_api.ts
@@ -30,7 +30,7 @@ export * from './lib/directives/text-highlight/text-highlight.directive';
 export * from './lib/directives/text-selection/text-selection.directive';
 export * from './lib/directives/template-outlet/template_outlet.directive';
 export * from './lib/directives/toggle/toggle.directive';
-export * from './lib/directives/tooltip/tooltip.directive';
+export * from './lib/directives/tooltip';
 export * from './lib/directives/date-time-editor/public_api';
 
 /**
