+++ src/app/tooltip/tooltip.sample.html
@@ -1,7 +1,7 @@
 <div class="column-container">
     <article class="column">
         <h4 class="sample-title">Simple tooltip</h4>
-        <igx-avatar #target="tooltipTarget" [igxTooltipTarget]="simple" [overlaySettings]="settings"
+        <igx-avatar #target="tooltipTarget" [igxTooltipTarget]="simple" [tooltip]="'Infragistics Inc. HQ'" [overlaySettings]="settings"
                     src="assets/images/avatar/10.jpg" size="medium" [roundShape]="true"
                     (onTooltipShow)="showing()"
                     (onTooltipHide)="hiding()"
@@ -10,8 +10,21 @@
 
         <div class="custom-width" #simple="tooltip" igxTooltip id="myid">
             Her name is Toola Tipa
-        </div>
+        </div>            
     </article>
+
+    <article class="column">
+        <h4 class="sample-title">Tooltip input</h4>
+        <div class="sample-tooltip-input">
+            <igx-avatar class="bottomMargin" src="assets/images/avatar/10.jpg" size="medium" [roundShape]="true" igxTooltipTarget [tooltip]="'Her name is Toola Tipa'">
+            </igx-avatar>
+
+            <igx-icon igxTooltipTarget [tooltip]="'Infragistics Inc. HQ'">
+                info
+            </igx-icon>
+        </div>      
+    </article>
+
     <article class="column">
         <h4 class="sample-title">Simple tooltip settings</h4>
         <div>
@@ -39,7 +52,7 @@
             <igx-card>
                 <igx-card-content style="padding: 0px;">
                     <div class="location">
-                        <igx-icon class="cursorHover" [color]="'rgb(0,153,255)'" [igxTooltipTarget]="locationTooltip">location_on</igx-icon>
+                        <igx-icon class="cursorHover" [style.color]="'rgb(0,153,255)'" [igxTooltipTarget]="locationTooltip">location_on</igx-icon>
 
                         <div class="locationTooltip" #locationTooltip="tooltip" igxTooltip>
                             <div class="locationTooltipContent">
@@ -90,5 +103,5 @@
                 </igx-column>
             </igx-grid>
         </div>
-    </article>
+    </article>    
 </div>
